@@ -1,6 +1,6 @@
 import logging
 import time
-from typing import Any, Dict, List
+from typing import Any
 
 import bleach
 
@@ -8,7 +8,7 @@ import bleach
 logger = logging.getLogger(__name__)
 
 # Rate limiting configuration
-RATE_LIMIT_WINDOWS: Dict[str, int] = {
+RATE_LIMIT_WINDOWS: dict[str, int] = {
     "1s": 1,  # 1 second window
     "1m": 60,  # 1 minute window
     "1h": 3600,  # 1 hour window
@@ -19,7 +19,7 @@ class RateLimiter:
     """Rate limiter implementation using sliding window."""
 
     def __init__(self):
-        self.requests: Dict[str, List[float]] = {}
+        self.requests: dict[str, list[float]] = {}
 
     def is_rate_limited(self, key: str, max_requests: int, window: float) -> bool:
         """Check if a key is rate limited.
@@ -39,9 +39,7 @@ class RateLimiter:
             self.requests[key] = []
 
         # Remove old requests outside window
-        self.requests[key] = [
-            req_time for req_time in self.requests[key] if now - req_time <= window
-        ]
+        self.requests[key] = [req_time for req_time in self.requests[key] if now - req_time <= window]
 
         # Check if rate limited
         if len(self.requests[key]) >= max_requests:
