@@ -1,7 +1,6 @@
 """Pydantic schemas for desktop configurations."""
 
 from datetime import datetime
-from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
@@ -10,11 +9,9 @@ class DesktopConfigurationBase(BaseModel):
     """Base schema for desktop configurations."""
 
     name: str = Field(..., description="Name of the desktop configuration")
-    description: Optional[str] = Field(None, description="Description of the desktop configuration")
+    description: str | None = Field(None, description="Description of the desktop configuration")
     image: str = Field(..., description="Docker image for the desktop environment")
-    is_public: bool = Field(
-        False, description="Whether the configuration is available to all users"
-    )
+    is_public: bool = Field(False, description="Whether the configuration is available to all users")
     min_cpu: int = Field(1, description="Minimum number of CPU cores")
     max_cpu: int = Field(4, description="Maximum number of CPU cores")
     min_ram: str = Field("4096Mi", description="Minimum RAM allocation (format: number + Mi/Gi)")
@@ -30,20 +27,14 @@ class DesktopConfigurationCreate(DesktopConfigurationBase):
 class DesktopConfigurationUpdate(BaseModel):
     """Schema for updating desktop configurations."""
 
-    name: Optional[str] = Field(None, description="Name of the desktop configuration")
-    description: Optional[str] = Field(None, description="Description of the desktop configuration")
-    image: Optional[str] = Field(None, description="Docker image for the desktop environment")
-    is_public: Optional[bool] = Field(
-        None, description="Whether the configuration is available to all users"
-    )
-    min_cpu: Optional[int] = Field(None, description="Minimum number of CPU cores")
-    max_cpu: Optional[int] = Field(None, description="Maximum number of CPU cores")
-    min_ram: Optional[str] = Field(
-        None, description="Minimum RAM allocation (format: number + Mi/Gi)"
-    )
-    max_ram: Optional[str] = Field(
-        None, description="Maximum RAM allocation (format: number + Mi/Gi)"
-    )
+    name: str | None = Field(None, description="Name of the desktop configuration")
+    description: str | None = Field(None, description="Description of the desktop configuration")
+    image: str | None = Field(None, description="Docker image for the desktop environment")
+    is_public: bool | None = Field(None, description="Whether the configuration is available to all users")
+    min_cpu: int | None = Field(None, description="Minimum number of CPU cores")
+    max_cpu: int | None = Field(None, description="Maximum number of CPU cores")
+    min_ram: str | None = Field(None, description="Minimum RAM allocation (format: number + Mi/Gi)")
+    max_ram: str | None = Field(None, description="Maximum RAM allocation (format: number + Mi/Gi)")
 
 
 class DesktopConfigurationInDB(DesktopConfigurationBase):
@@ -87,6 +78,4 @@ class DesktopConfigurationAccessInDB(DesktopConfigurationAccessBase):
 class DesktopConfigurationWithAccess(DesktopConfigurationInDB):
     """Schema for desktop configurations with access information."""
 
-    allowed_users: List[str] = Field(
-        [], description="Usernames of users with access to this configuration"
-    )
+    allowed_users: list[str] = Field([], description="Usernames of users with access to this configuration")
