@@ -11,9 +11,7 @@ from desktop_manager.core.exceptions import APIError
 def test_create_user_validation_error(client, mock_db_client):
     """Test user creation with validation error."""
     # Make request with validation error (missing username)
-    response = client.post(
-        "/test_create_user", json={"email": "test@example.com"}, content_type="application/json"
-    )
+    response = client.post("/test_create_user", json={"email": "test@example.com"}, content_type="application/json")
 
     # Should be BAD_REQUEST
     assert response.status_code == HTTPStatus.BAD_REQUEST
@@ -41,9 +39,7 @@ def test_create_user_with_custom_validation_error(client, mock_db_client):
 def test_create_user_database_error(client, mock_db_client):
     """Test user creation with database error."""
     # Configure mock to raise database error
-    mock_db_client.execute_query.side_effect = APIError(
-        "Database error", HTTPStatus.INTERNAL_SERVER_ERROR
-    )
+    mock_db_client.execute_query.side_effect = APIError("Database error", HTTPStatus.INTERNAL_SERVER_ERROR)
 
     # Make request
     response = client.post(
@@ -74,9 +70,7 @@ def test_check_user_without_parameters(client):
 def test_check_user_database_error(client, mock_db_client):
     """Test checking user existence with database error."""
     # Configure mock to raise database error
-    mock_db_client.execute_query.side_effect = APIError(
-        "Database error", HTTPStatus.INTERNAL_SERVER_ERROR
-    )
+    mock_db_client.execute_query.side_effect = APIError("Database error", HTTPStatus.INTERNAL_SERVER_ERROR)
 
     # Make request with username parameter
     response = client.get("/test_check_user?username=testuser")
